Environment setup
Added Planes, ports where viruses will enter the game. The colors i used standard colors, i don't know why they are being displayed, i also added a navmesh ai to cube with names as node, they all have a tag of (Path).
Just trying to give an image of the game am getting in my head.
If the direction is this our artist will have to develop motherboard pathways and some important components we might like on on the board